+++ src/main/java/wooteco/subway/controller/LineController.java
@@ -16,7 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.dto.LineRequest;
 import wooteco.subway.dto.LineResponse;
 import wooteco.subway.dto.SectionRequest;
-import wooteco.subway.dto.SectionResponse;
 import wooteco.subway.service.LineService;
 
 @RestController
@@ -60,7 +59,8 @@ public class LineController {
     }
 
     @PostMapping(value = "{id}/sections")
-    public ResponseEntity<Void> createSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
+    public ResponseEntity<Void> createSection(@PathVariable Long id,
+        @RequestBody SectionRequest sectionRequest) {
         lineService.createSection(id, sectionRequest);
         return ResponseEntity.ok().build();
     }
