+++ src/test/java/wooteco/subway/acceptance/station/StationAcceptanceTest.java
@@ -15,7 +15,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.jdbc.Sql;
 import wooteco.subway.acceptance.AcceptanceTest;
 import wooteco.subway.dto.StationResponse;
